rename and go,ec fix
