+++ uniprotkb-rest/src/main/java/org/uniprot/api/uniprotkb/controller/GroupByController.java
@@ -1,18 +1,17 @@
 package org.uniprot.api.uniprotkb.controller;
 
-import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-import static org.uniprot.api.uniprotkb.controller.UniProtKBController.UNIPROTKB_RESOURCE;
-
-import org.springframework.http.ResponseEntity;
-import org.uniprot.api.uniprotkb.groupby.model.GroupByResult;
-
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import org.springframework.http.ResponseEntity;
+import org.uniprot.api.uniprotkb.groupby.model.GroupByResult;
 
-public abstract class UniProtKBGroupByController {
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+import static org.uniprot.api.uniprotkb.controller.UniProtKBController.UNIPROTKB_RESOURCE;
+
+public abstract class GroupByController {
     static final String GROUP_BY_RESOURCE = UNIPROTKB_RESOURCE + "/groups";
 
     @Tag(name = "uniprotkbgroup")
