+++ uniprotkb-rest/src/main/java/org/uniprot/api/uniprotkb/controller/GroupByTaxonomyController.java
@@ -9,21 +9,21 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.uniprot.api.uniprotkb.groupby.model.GroupByResult;
-import org.uniprot.api.uniprotkb.groupby.service.UniProtKBGroupByKeywordService;
+import org.uniprot.api.uniprotkb.groupby.service.GroupByTaxonomyService;
 
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-import static org.uniprot.api.uniprotkb.controller.UniProtKBGroupByKeywordController.GROUP_BY_KEYWORD_RESOURCE;
+import static org.uniprot.api.uniprotkb.controller.GroupByTaxonomyController.GROUP_BY_TAXONOMY_RESOURCE;
 
-@RequestMapping(value = GROUP_BY_KEYWORD_RESOURCE)
+@RequestMapping(value = GROUP_BY_TAXONOMY_RESOURCE)
 @RestController
-public class UniProtKBGroupByKeywordController extends UniProtKBGroupByController {
-    static final String GROUP_BY_KEYWORD_RESOURCE = GROUP_BY_RESOURCE + "/keyword";
-    private final UniProtKBGroupByKeywordService uniProtKBGroupByKeywordService;
+public class GroupByTaxonomyController extends GroupByController {
+    static final String GROUP_BY_TAXONOMY_RESOURCE = GROUP_BY_RESOURCE + "/taxonomy";
+    private final GroupByTaxonomyService uniProtKBGroupByTaxonomyService;
 
     @Autowired
-    public UniProtKBGroupByKeywordController(
-            UniProtKBGroupByKeywordService uniProtKBGroupByKeywordService) {
-        this.uniProtKBGroupByKeywordService = uniProtKBGroupByKeywordService;
+    public GroupByTaxonomyController(
+            GroupByTaxonomyService uniProtKBGroupByTaxonomyService) {
+        this.uniProtKBGroupByTaxonomyService = uniProtKBGroupByTaxonomyService;
     }
 
     @Override
@@ -38,6 +38,6 @@ public class UniProtKBGroupByKeywordController extends UniProtKBGroupByControlle
                     @RequestParam(value = "parent", required = false)
                     String parent) {
         return new ResponseEntity<>(
-                uniProtKBGroupByKeywordService.getGroupByResult(query, parent), HttpStatus.OK);
+                uniProtKBGroupByTaxonomyService.getGroupByResult(query, parent), HttpStatus.OK);
     }
 }
