+++ uniprotkb-rest/src/main/java/org/uniprot/api/uniprotkb/groupby/GroupByServiceConfig.java
@@ -10,7 +10,7 @@ import org.uniprot.cv.ec.ECRepoFactory;
 public class GroupByServiceConfig {
     private final String ecDirectory;
 
-    public GroupByServiceConfig(@Value("${groupby.solr.ec.dir}") String ecDirectory) {
+    public GroupByServiceConfig(@Value("${groupby.ec.dir}") String ecDirectory) {
         this.ecDirectory = ecDirectory;
     }
