+++ uniprotkb-rest/src/main/resources/application.properties
@@ -95,7 +95,15 @@ uniprot.rdf.streamer.batchSize=200
 uniprot.rdf.streamer.MaxRetries=3
 uniprot.rdf.streamer.retryDelayMillis=1000
 
-groupby.solr.ec.dir=/my/ec/dir
+groupby.ec.dir=ec/
+groupby.go.quickgo.url=https://www.ebi.ac.uk/QuickGO/services/ontology/go/terms/
+groupby.go.roots[0].id=GO:0008150
+groupby.go.roots[0].name=biological_process
+groupby.go.roots[1].id=GO:0005575
+groupby.go.roots[1].name=cellular_component
+groupby.go.roots[2].id=GO:0003674
+groupby.go.roots[2].name=molecular_function
+
 
 ############################### Term properties ###############################
 terms.fields=cc_catalytic_activity,lit_author,cc_disease,gene,go,keyword,family,protein_name,strain,taxonomy_name,organism_name,tissue
