+++ uniprotkb-rest/src/test/java/org/uniprot/api/uniprotkb/controller/GroupByControllerIT.java
@@ -2,6 +2,8 @@ package org.uniprot.api.uniprotkb.controller;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.test.web.servlet.MockMvc;
+import org.uniprot.store.indexer.DataStoreManager;
+import org.uniprot.store.search.document.Document;
 
 import static org.hamcrest.Matchers.containsStringIgnoringCase;
 import static org.hamcrest.core.Is.is;
@@ -9,7 +11,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.log;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
-public abstract class UniProtKBGroupByControllerIT {
+public abstract class GroupByControllerIT {
     protected static final String EMPTY_PARENT = "";
     private static final String INVALID_ORGANISM_ID = "36";
 
@@ -49,7 +51,18 @@ public abstract class UniProtKBGroupByControllerIT {
                                                 "query is a required parameter")));
     }
 
+    protected void save(DataStoreManager.StoreType type, Document doc) {
+        getDataStoreManager().saveDocs(type, doc);
+    }
+
+    protected void save(Document doc) {
+        getDataStoreManager().saveDocs(DataStoreManager.StoreType.UNIPROT, doc);
+    }
+
+    protected abstract DataStoreManager getDataStoreManager();
+
     protected abstract MockMvc getMockMvc();
+
     protected abstract String getPath();
 
     protected abstract void prepareSingleRootNodeWithNoChildren() throws Exception;
